+++ internal/bloblang/query/methods.go
@@ -129,28 +129,27 @@ var _ = RegisterMethod(
 		NewExampleSpec("",
 			`root.doc.id = this.thing.id.string().catch(uuid_v4())`,
 		),
+		NewExampleSpec("When the input document is not structured attempting to reference structured fields with `this` will result in an error. Therefore, a convenient way to delete non-structured data is with a catch.",
+			`root = this.catch(deleted())`,
+			`{"doc":{"foo":"bar"}}`,
+			`{"doc":{"foo":"bar"}}`,
+			`not structured data`,
+			`<Message deleted>`,
+		),
 	),
 	false, catchMethod,
 	ExpectNArgs(1),
 )
 
 func catchMethod(fn Function, args ...interface{}) (Function, error) {
-	var catchFn Function
-	switch t := args[0].(type) {
-	case uint64, int64, float64, json.Number, string, []byte, bool, []interface{}, map[string]interface{}:
-		catchFn = NewLiteralFunction(t)
-	case Function:
-		catchFn = t
-	default:
-		return nil, fmt.Errorf("expected query or literal argument, received %T", args[0])
-	}
+	catchVal := args[0]
 	return ClosureFunction(func(ctx FunctionContext) (interface{}, error) {
 		res, err := fn.Exec(ctx)
 		if err != nil {
-			res, err = catchFn.Exec(ctx)
+			return catchVal, nil
 		}
 		return res, err
-	}, aggregateTargetPaths(fn, catchFn)), nil
+	}, fn.QueryTargets), nil
 }
 
 //------------------------------------------------------------------------------